feat: implement MoneyInput component for handling VND input with formatting and validation

+++ apps/crm-web/src/app/(dashboard)/quotes/schema.ts
@@ -9,7 +9,9 @@ const quoteItemSchema = z.object({
   description: z.string().min(1, "Nhập hạng mục"),
   unit: z.string().optional(),
   quantity: z.number().min(0),
-  unit_price: z.number().min(0),
+  // Whole đồng only — VND has no fractional unit and the column is BigInt, so a
+  // decimal here was silently dropped server-side.
+  unit_price: z.number().int().min(0),
 });
 
 export const createQuoteSchema = z.object({
